refactor(inspector): flatten component folder structure for fields
