+++ src/features/inspector/components/coverage-section/CoverageField.tsx
