+++ src/features/inspector/components/env-section/ObstacleSumField.tsx
