+++ src/features/inspector/components/object-section/CategoryField.tsx
