+++ src/features/inspector/components/object-section/GrossAreaField.tsx
