+++ src/features/inspector/components/object-section/NetAreaField.tsx
