+++ src/features/inspector/components/object-section/VertexSumField.tsx
