+++ src/layouts/workbench-window/inspector-panel/coverage-section/CoverageSection.tsx
@@ -1,7 +1,7 @@
 import InspectorPanelSection from "@/components/inspector-panel/InspectorPanelSection";
-import CoverageField from "@/features/inspector/components/coverage-section/CoverageField/CoverageField";
-import OverlapField from "@/features/inspector/components/coverage-section/OverlapField/OverlapField";
-import TurnSumField from "@/features/inspector/components/coverage-section/TurnSumField/TurnSumField";
+import CoverageField from "@/features/inspector/components/coverage-section/CoverageField";
+import OverlapField from "@/features/inspector/components/coverage-section/OverlapField";
+import TurnSumField from "@/features/inspector/components/coverage-section/TurnSumField";
 
 export default function CoverageSection() {
   return (
