+++ src/layouts/workbench-window/inspector-panel/env-section/EnvSection.tsx
@@ -1,6 +1,6 @@
 import InspectorPanelSection from "@/components/inspector-panel/InspectorPanelSection";
-import ObstacleSumField from "@/features/inspector/components/env-section/ObstacleSumField/ObstacleSumField";
-import ZoneSumField from "@/features/inspector/components/env-section/ZoneSumField/ZoneSumField";
+import ObstacleSumField from "@/features/inspector/components/env-section/ObstacleSumField";
+import ZoneSumField from "@/features/inspector/components/env-section/ZoneSumField";
 
 export default function EnvSection() {
   return (
