+++ src/layouts/workbench-window/inspector-panel/object-section/ObjectSection.tsx
@@ -1,9 +1,9 @@
 import InspectorPanelSection from "@/components/inspector-panel/InspectorPanelSection";
-import TypeField from "@/features/inspector/components/object-section/TypeField/TypeField";
-import CategoryField from "@/features/inspector/components/object-section/CategoryField/CategoryField";
-import GrossAreaField from "@/features/inspector/components/object-section/GrossAreaField/GrossAreaField";
-import NetAreaField from "@/features/inspector/components/object-section/NetAreaField/NetAreaField";
-import VertexSumField from "@/features/inspector/components/object-section/VertexSumField/VertexSumField";
+import TypeField from "@/features/inspector/components/object-section/TypeField";
+import CategoryField from "@/features/inspector/components/object-section/CategoryField";
+import GrossAreaField from "@/features/inspector/components/object-section/GrossAreaField";
+import NetAreaField from "@/features/inspector/components/object-section/NetAreaField";
+import VertexSumField from "@/features/inspector/components/object-section/VertexSumField";
 
 export default function ObjectSection() {
   return (
